+++ src/set_iso_year.js
@@ -1,5 +1,9 @@
 var parse = require('./parse')
+var getTimeInMilliseconds = require('./get_time_in_milliseconds')
 var startOfISOYear = require('./start_of_iso_year')
+var differenceInDays = require('./difference_in_days')
+
+var MILLISECONDS_IN_DAY = 86400000
 
 /**
  * Sets ISO week-numbering year to given date,
@@ -13,9 +17,13 @@ var startOfISOYear = require('./start_of_iso_year')
  */
 var setISOYear = function(dirtyDate, isoYear) {
   var date = parse(dirtyDate)
-  var startOfThisYear = startOfISOYear(date)
-  var diff = date.getTime() - startOfThisYear.getTime()
-  date.setTime(startOfISOYear(new Date(isoYear, 0, 4)).getTime() + diff)
+
+  var time = getTimeInMilliseconds(date)
+  var diff = differenceInDays(date, startOfISOYear(date))
+
+  date = startOfISOYear(new Date(isoYear, 0, 4))
+  date.setDate(date.getDate() + diff)
+  date.setHours(0, 0, 0, time)
   return date
 }