+++ src/add_months.js
@@ -1,5 +1,5 @@
 var parse = require('./parse')
-var getTimeInMilliseconds = require('./get_time_in_milliseconds')
+var getTimeSinceMidnight = require('./get_time_since_midnight')
 var getDaysInMonth = require('./get_days_in_month')
 
 /**
@@ -10,7 +10,7 @@ var getDaysInMonth = require('./get_days_in_month')
  */
 var addMonths = function(dirtyDate, amount) {
   var date = parse(dirtyDate)
-  var time = getTimeInMilliseconds(date)
+  var time = getTimeSinceMidnight(date)
   var desiredMonth = date.getMonth() + amount
   var daysInMonth = getDaysInMonth(new Date(date.getFullYear(), desiredMonth, 1))