+++ src/date_fns.js
@@ -34,7 +34,7 @@ var dateFns = {
   getMonth: require('./get_month'),
   getQuarter: require('./get_quarter'),
   getSeconds: require('./get_seconds'),
-  getTimeInMilliseconds: require('./get_time_in_milliseconds'),
+  getTimeSinceMidnight: require('./get_time_since_midnight'),
   getYear: require('./get_year'),
   isAfter: require('./is_after'),
   isBefore: require('./is_before'),