+++ CHANGELOG.md
@@ -9,7 +9,8 @@
   - `setISOWeek` now keeps time from original date
   - Reuse `getDaysInMonth` inside of `addMonths`
 * Bug fixes:
-  - Fix a lot of DST-related bugs
+  - Fix a lot of bugs appearing when date is modifying into other time zone (e.g., when adding months and original date is in DST but new date is not)
+  - Prevent instances of Date to lose milliseconds value when passed to `parse` in IE10
 
 ## v0.16.0 (1 Sep 2015)